forskellige ting

+++ VideoCoding/quantize.h
@@ -7,7 +7,7 @@ class Quantize
 {
 public:
 	static std::vector<char> quant(std::vector<int_fast16_t>);
-	static void setQuality(int_fast16_t);
+	static void setQuality(uint_fast8_t);
 private:
 	// static uint_fast8_t quality;
 	const static std::array<int_fast16_t, 64> lumTableOrig;
